% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arepa.R
\name{get_AQS_data_annual}
\alias{get_AQS_data_annual}
\title{Download annual AQS datasets}
\usage{
get_AQS_data_annual(year = 1990:2013)
}
\arguments{
\item{\code{year}}{A year or list of years (from 1990 to 2018)}
}
\value{
This function is used for its side-effects: it downloads annual AQS datasets
from \url{http://aqsdr1.epa.gov/aqsweb/aqstmp/airdata/download_files.html}
in local folder 'Data_AQS'.
}
\description{
Download annual AQS datasets
}
\examples{
get_AQS_data_annual(2000:2002)
}
