% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/arepa.R
\name{get_AQS_data_daily}
\alias{get_AQS_data_daily}
\title{Download daily AQS datasets}
\usage{
get_AQS_data_daily(parameter = 88101, year = 2002:2004)
}
\arguments{
\item{\code{parameter}}{Parameter code (e.g, 88101 for PM 2.5 FRM)}

\item{\code{year}}{A year or list of years (from 1990 to 2014)}
}
\value{
This function is used for its side-effects: it downloads annual AQS datasets
from \url{http://aqsdr1.epa.gov/aqsweb/aqstmp/airdata/download_files.html}
in local folder 'Data_AQS'.
}
\description{
Download daily AQS datasets
}
\examples{
get_AQS_data_daily(2000:2002)
}

