% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/arepa.R
\name{get_zip_codes}
\alias{get_zip_codes}
\title{Retrieve zip codes}
\usage{
get_zip_codes()
}
\arguments{
\item{x}{Any R object.}

\item{y}{Any R object.}
}
\value{
A list with first element \code{x} and second element \code{y}.
}
\description{
Retrieve zip codes
}
\examples{
polenta(10, 1)
polenta("10", 1:5)
}

