% Generated by roxygen2 (4.0.2): do not edit by hand
\name{polenta}
\alias{polenta}
\title{Returns a list of two elements.}
\usage{
polenta(x, y)
}
\arguments{
\item{x}{Any R object.}

\item{y}{Any R object.}
}
\value{
A list with first element \code{x} and second element \code{y}.
}
\description{
Returns a list of two elements.
}
\examples{
polenta(10, 1)
polenta("10", 1:5)
}

