% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/arepa.R
\name{spatial_link_index}
\alias{spatial_link_index}
\title{Spatial index distance matrix}
\usage{
spatial_link_index(data1, lat1, lon1, id1, data2, lat2, lon2, id2,
  within = NULL)
}
\arguments{
\item{\code{data1}}{Dataset 1}

\item{\code{lat1}}{Latitude variable in dataset 1}

\item{\code{lon1}}{Longitude variable in dataset 1}

\item{\code{id1}}{Unique key in dataset 1}

\item{\code{data2}}{Dataset 1}

\item{\code{lat2}}{Latitude variable in dataset 2}

\item{\code{lon2}}{Longitude variable in dataset 2}

\item{\code{id2}}{Unique key in dataset 2}

\item{\code{within}}{Link \code{within} kilometers}
}
\value{
This function returns a data table of pairwise distances \code{within}
kilometers with column names \code{id1}, \code{id2} and \code{Distance}.
}
\description{
Spatial index distance matrix
}
\examples{
years <- 2000:2002 # possible range 1990-2014
within_km <- 9.656 # 6 miles
parameter_code <- 81102 # for PM 10
##----- Get PM10 annual data and set unique Monitor key
# Download AQS data in Data_AQS/all.
# get_AQS_data_annual(years)
PM <- load_annual_average(years)
PM <- subset(PM, Parameter.Code == parameter_code)
##----- Zip codes
ZIP <- get_zip_codes()
##----- Spatial index to join monitors and Zip codes.  May take a couple of minutes
Link_PM_Zip_Index <- spatial_link_index(PM, "Latitude", "Longitude", "Monitor",
                                        ZIP, "Latitude.zip", "Longitude.zip", "zip",
                                        within = within_km)
}

