% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/arepa.R
\name{subset_monitors}
\alias{subset_monitors}
\title{Create a subset of AQS monitors}
\usage{
subset_monitors(MONITORS, parameter_code, observation_percent,
  monitor_info_file = "monitor_list.csv")
}
\arguments{
\item{\code{MONITORS}}{An AQS data table obtained created with \code{\link{load_annual_average}}}

\item{\code{parameter_code}}{A parameter code}

\item{\code{observation_percent}}{The required valid observation percent, e.g., 67}

\item{\code{monitor_info_file}}{The path of the file containing extra information
       about monitors: "monitor_list.csv"}
}
\value{
This function returns a data table of annual AQS datasets previously
downloaded in local folder 'Data_AQS' with function \code{\link{get_AQS_data_annual}}.
}
\description{
Create a subset of AQS monitors
}
\examples{
get_AQS_data_annual(2000:2002)
list_AQS <- lapply(2000:2002, load_annual_average)
AQS <- rbindlist(list_AQS)
sub_AQS <- subset_monitors(AQS, 81102, 67, "~/Google Drive/ARP/Arepa/monitor_list.csv")
}

