% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/arepa.R
\name{get_zip_codes}
\alias{get_zip_codes}
\title{Retrieve zip codes}
\usage{
get_zip_codes()
}
\value{
A data table with columns \code{city}, \code{state}, \code{latitude}
and \code{longitude}.
}
\description{
Retrieve zip codes
}
\examples{
get_zip_codes()
}

