% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arepa.R
\name{load_annual_average}
\alias{load_annual_average}
\title{Load annual AQS datasets}
\usage{
load_annual_average(year)
}
\arguments{
\item{\code{year}}{A year or list of years (from 1990 to 2018)}
}
\value{
This function returns a data table of annual AQS datasets previously
downloaded in local folder 'Data_AQS' with function \code{\link{get_AQS_data_annual}}. A unique
monitor key \code{Monitor} is created.
}
\description{
Load annual AQS datasets
}
\examples{
get_AQS_data_annual(2000:2002)
AQS <- load_annual_average(2000:2002)
}
