% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arepa.R
\name{load_daily_data}
\alias{load_daily_data}
\title{Load daily AQS datasets}
\usage{
load_daily_data(parameter = 88101, year = 2002:2004)
}
\arguments{
\item{\code{parameter}}{Parameter code (e.g, 88101 for PM 2.5 FRM)}

\item{\code{year}}{A year or list of years (from 1990 to 2014)}
}
\value{
This function returns a data table of annual AQS datasets previously
downloaded in local folder 'Data_AQS' with function \code{\link{get_AQS_data_annual}}. A unique
monitor key \code{Monitor} is created.
}
\description{
Load daily AQS datasets
}
\examples{
get_AQS_data_annual(2000:2002)
AQS <- load_daily_data(2000:2002)
}
